requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
beautifulsoup4==4.12.3
    # via -r requirements.in
black==24.1.1
    # via -r requirements.in
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   djlint
colorama==0.4.6
    # via djlint
cssbeautifier==1.14.11
    # via djlint
decorator==5.1.1
    # via ipython
django==5.0.1
    # via
    #   -r requirements.in
    #   django-auto-prefetch
    #   django-hyperview
django-auto-prefetch==1.8.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
django-hyperview==0.0.2
    # via -r requirements.in
djlint==1.34.1
    # via -r requirements.in
editorconfig==0.12.3
    # via
    #   cssbeautifier
    #   jsbeautifier
exceptiongroup==1.2.0
    # via ipython
executing==2.0.1
    # via stack-data
gunicorn==21.2.0
    # via -r requirements.in
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
idna==3.6
    # via requests
ipython==8.21.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jsbeautifier==1.14.11
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.14
    # via djlint
matplotlib-inline==0.1.6
    # via ipython
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   gunicorn
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   djlint
pexpect==4.9.0
    # via ipython
platformdirs==4.2.0
    # via black
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.17.2
    # via ipython
pyyaml==6.0.1
    # via djlint
regex==2023.12.25
    # via djlint
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   cssbeautifier
    #   jsbeautifier
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
tomli==2.0.1
    # via
    #   black
    #   djlint
tqdm==4.66.1
    # via djlint
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   asgiref
    #   black
urllib3==2.2.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
